+++ tencentcloud/__init__.py
@@ -14,4 +14,4 @@
 # limitations under the License.
 
 
-__version__ = '3.0.734'
+__version__ = '3.0.735'